src/main/java/bg/exam/animal4me/service/Animal4meUserDetailsService.java
package bg.exam.animal4me.service;

import bg.exam.animal4me.model.entity.UserEntity;
import bg.exam.animal4me.model.entity.UserRoleEntity;
import bg.exam.animal4me.model.user.Animal4meUserDetails;
import bg.exam.animal4me.repository.UserRepository;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;

public class Animal4meUserDetailsService implements UserDetailsService {

    private  UserRepository userRepository;

    public Animal4meUserDetailsService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String username)
            throws UsernameNotFoundException {
        return userRepository.
                findByEmail(username).
                map(this::map).
                orElseThrow(() -> new UsernameNotFoundException("User with email " + username + " not found!"));
    }

    private UserDetails map(UserEntity userEntity) {

        return new Animal4meUserDetails(
                userEntity.getPassword(),
                userEntity.getEmail(),
                userEntity.getFirstName(),
                userEntity.getLastName(),
                userEntity.getTelephone()
                , userEntity.
                getUserRoles().
                stream().
                map(this::map).
                toList()
        );
    }

    private GrantedAuthority map(UserRoleEntity userRole) {
        return new SimpleGrantedAuthority("ROLE_" +
                userRole.
                        getUserRole().name());
    }
}